performance optimization.
Pattern is thread-safe, so no point in recompiling in every parse.

As for DateFormat, where the performance starts to matter is in a large instance, which normally has a lot of executors and concurrent activities going on. So you gain more by concurrent access to them than by a little memory saving you get by sharing instances. So it's better to simply create those formatters on stack and throw them away at the end.